# dist/openapi.yaml
openapi: 3.0.3
info:
  title: Recipes API
  description: API for recipes project.
  version: 1.0.0
servers:
  - description: Localhost API
    url: http://localhost
  # - description: Partner's mock server
  #   url: https://virtserver.swaggerhub.com/AmitDvash/WebDevelopmentCourse/1.0.0
...
paths:
...
  /users/markwatched/{recipeId}:
    post:
      tags:
        - Users
      summary: Mark a recipe as watched
      description: Marks the specified recipe as watched by the logged-in user.
      parameters:
        - in: path
          name: recipeId
          required: true
          schema:
            type: string
            description: Recipe unique ID
      responses:
        "200":
          description: Recipe marked as watched
        "401":
          description: User not logged in
  /users/deleteWatchedRecipes:
    delete:
      tags:
        - Users
      summary: Delete all watched recipes
      description: Removes all the watched recipes for the logged-in user.
      responses:
        "200":
          description: All watched recipes deleted
        "401":
          description: User not logged in
...
  /users/favorites:
    get:
      tags:
        - Users
      summary: Return user's favorite recipes
      description: Returns the favorite recipes saved by the logged-in user.
      responses:
        "200":
          description: A list of all user's favorite recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/previewRecipe"
        "401":
          description: Unauthorized

    post:
      tags:
        - Users
      summary: Save a recipe in the user's favorites list
      description: Adds the chosen recipe to the favorites list of the logged-in user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeId:
                  type: integer
                  description: The recipe's ID to save
                  example: 716429
      responses:
        "200":
          description: Recipe successfully saved in the favorites list
        "400":
          description: Recipe does not exist
        "401":
          description: Unauthorized

    delete:
      tags:
        - Users
      summary: Remove a recipe from the user's favorites list
      description: Removes the chosen recipe from the favorites list of the logged-in user.
      parameters:
        - in: query
          name: recipeId
          required: true
          schema:
            type: integer
            description: The recipe's ID to remove
            example: 716429
      responses:
        "200":
          description: Recipe successfully removed from the favorites list
        "401":
          description: Unauthorized
  /users/myRecipes:
    get:
      tags:
        - Users
      summary: Return user's private recipes
      description: Returns a JSON object containing preview details about all the private recipes of the logged-in user.
      responses:
        "200":
          description: List of all user's private recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrivateRecipe"
        "401":
          description: Unauthorized

    post:
      tags:
        - Users
      summary: Add a new recipe to the user's private recipes
      description: Adds a new recipe to the private list of the logged-in user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFullPrivateRecipe"
      responses:
        "201":
          description: The private recipe was successfully created
        "400":
          description: Missing required parameters
        "401":
          description: Unauthorized
  /users/myRecipes/{recipeId}:
    get:
      tags:
        - Users
      summary: Return full details of a private recipe
      description: Returns a JSON object containing the full details about this private recipe.
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
            description: Private recipe unique ID
      responses:
        "200":
          description: Private recipe full details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullPrivateRecipe"
        "401":
          description: Unauthorized
        "404":
          description: Private recipe with the given ID does not exist
    delete:
      tags:
        - Users
      summary: Delete a private recipe
      description: Deletes a specific private recipe that belongs to the logged-in user.
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
            description: Private recipe unique ID
      responses:
        "200":
          description: Private recipe deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Recipe deleted successfully"
                  success:
                    type: boolean
                    example: true
        "400":
          description: Invalid recipe ID format
        "401":
          description: Unauthorized
        "404":
          description: Private recipe not found or not owned by user
        "500":
          description: Failed to delete private recipe
  /users/lastWatchedRecipes:
    get:
      tags:
        - Users
      summary: Return the user's last 3 watched recipes
      description: Returns preview details about the last 3 watched recipes of the logged-in user.
      responses:
        "200":
          description: Preview details of the last 3 watched recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/previewRecipe"
        "401":
          description: Unauthorized
  /users/allWatchedRecipes:
    get:
      tags:
        - Users
      summary: Return the user's watched recipes
      description: Returns preview details about the watched recipes of the logged-in user.
      responses:
        "200":
          description: Preview details of the watched recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/previewRecipe"
        "401":
          description: Unauthorized
  /users/lastSearch:
    get:
      tags:
        - Users
      summary: Return the user's last search results
      description:
        Returns a list of all the matching recipes returned in the last search of the logged-in user.
        If the user disconnects from the site or connects from another browser, the last search results will not be shown.
      responses:
        "200":
          description: User's last search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/previewRecipe"
        "401":
          description: Unauthorized
  /users/familyRecipes:
    get:
      tags:
        - Users
      summary: Get all family recipes for the logged-in user
      description: Returns a list of all family recipes belonging to the user.
      responses:
        "200":
          description: A list of family recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/previewFamilyRecipe"
        "204":
          description: Fewer than 3 family recipes found
        "401":
          description: Unauthorized
    post:
      tags:
        - Users
      summary: Add a new family recipe
      description: Adds a new family recipe for the logged-in user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFamilyRecipe"
      responses:
        "201":
          description: Family recipe created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
                  recipe_id:
                    type: integer
        "400":
          description: Missing required fields
        "401":
          description: Unauthorized
        "500":
          description: Failed to create family recipe
  /users/familyRecipes/{recipeId}:
    get:
      tags:
        - Users
      summary: Get full details of a specific family recipe
      description: Returns the complete details of a family recipe including ingredients, instructions, and preparation information.
      parameters:
        - in: path
          name: recipeId
          required: true
          schema:
            type: string
          description: The ID of the family recipe to retrieve
      responses:
        "200":
          description: Family recipe details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullFamilyRecipe"
        "401":
          description: Unauthorized
        "404":
          description: Recipe not found or not owned by user
    delete:
      tags:
        - Users
      summary: Delete a family recipe
      description: Deletes a specific family recipe for the logged-in user.
      parameters:
        - in: path
          name: recipeId
          required: true
          schema:
            type: string
          description: The ID of the family recipe to delete
      responses:
        "200":
          description: Family recipe deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
        "400":
          description: Invalid recipe ID format
        "401":
          description: Unauthorized
        "404":
          description: Family recipe not found or not owned by user
        "500":
          description: Failed to delete family recipe
...